requirements.txt
annotated-types==0.7.0
anyio==4.8.0
autopep8==2.3.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
filetype==1.2.0
Flask==3.0.3
google-ai-generativelanguage==0.6.16
google-api-core==2.24.2
google-auth==2.38.0
google-genai==1.7.0
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.20
langchain-core==0.3.45
langchain-google-genai==2.1.0
langchain-text-splitters==0.3.6
langgraph==0.3.11
langgraph-checkpoint==2.0.20
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.57
langsmith==0.3.15
MarkupSafe==3.0.2
msgpack==1.1.0
orjson==3.10.15
packaging==24.2
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.39
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
websockets==15.0.1
Werkzeug==3.0.6
zstandard==0.23.0
